refactor: :recycle: pull out relabel_block into its own function
